Разделение создания фото на модули
