+++ index.html
@@ -7,7 +7,10 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <!-- <script type="text/javascript" src="/js/functions.js"></script> -->
-  <script type="text/javascript" src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
+  <script type="module" src="/js/util.js"></script>
+  <script type="module" src="/js/data.js"></script>
+
   <title>Кекстаграм</title>
 </head>
